updated tz/zoneinfo to version 2025a

+++ scripts/mktzdata.sh
@@ -8,7 +8,7 @@ ZIC=zic
 CP=cp
 RM=rm
 TZLIB="src/tools/tz"
-TZDIR="../../../dist/common/YAM/Resources/Zoneinfo/"
+TZDIR="../../../dist/common/YAM/Resources/Zoneinfo"
 
 if [ ! -e "${TZLIB}" ]; then
   echo "ERROR: You have to call this script from the top-level directory."
@@ -24,7 +24,7 @@ ${RM} -rf ${TZDIR}/*
 # use 'zic' to generate the zoneinfo files we require for yam
 TIMEZONES="africa antarctica asia australasia europe northamerica southamerica"
 for tz in ${TIMEZONES}; do
-  ${ZIC} -y yearistype.sh -d ${TZDIR} -v ${tz}
+  ${ZIC} -d ${TZDIR} -v ${tz}
 done
 
 # copy the .tab files because YAM will use them
@@ -34,5 +34,7 @@ ${CP} zone1970.tab ${TZDIR}/zone.tab
 # but yam will not use
 RMFILES="CET CST6CDT EET EST EST5EDT HST MET MST MST7MDT PST8PDT WET"
 for file in ${RMFILES}; do
-  ${RM} ${TZDIR}/${file}
+  if [ -f ${TZDIR}/${file} ]; then
+    ${RM} ${TZDIR}/${file}
+  fi
 done
